Add exemption for a kind of Google embedded videos
This is seen in Google Drive and in 3rd-party viewers like Invidious (a free
front-end for YouTube intended to remove tracking).

This is whitelisted because the URL for these videos is signed, and removing
any single parameter (even the IP) causes the video not to load. Disabling
this filter will break these sites.